# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.createListingRequest:
    properties:
      categoryId:
        type: string
      condition:
        type: string
      description:
        type: string
      price:
        type: integer
      title:
        type: string
    type: object
...
  handler.updateListingRequest:
    properties:
      categoryId:
        type: string
      condition:
        type: string
      description:
        type: string
      price:
        type: integer
      title:
        type: string
    type: object
...
  handler.updateStatusRequest:
    properties:
      status:
        type: string
    type: object
...
  models.Listing:
    properties:
      categoryId:
        type: string
      condition:
        $ref: '#/definitions/models.ListingCondition'
      createdAt:
        type: string
      deletedAt:
        description: nil = ยังไม่ถูกลบ
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ListingImage'
        type: array
      price:
        description: หน่วยเป็นบาท เก็บเป็นจำนวนเต็มเพื่อเลี่ยงปัญหา float
        type: integer
      seller:
        $ref: '#/definitions/models.PublicUser'
      sellerId:
        type: string
      status:
        $ref: '#/definitions/models.ListingStatus'
      suggestedPrice:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.ListingCondition:
    enum:
    - new
    - like_new
    - good
    - worn
    type: string
    x-enum-varnames:
    - ConditionNew
    - ConditionLikeNew
    - ConditionGood
    - ConditionWorn
  models.ListingImage:
    properties:
      id:
        type: string
      listingId:
        type: string
      sortOrder:
        type: integer
      url:
        type: string
    type: object
  models.ListingStatus:
    enum:
    - available
    - reserved
    - sold
    type: string
    x-enum-varnames:
    - StatusAvailable
    - StatusReserved
    - StatusSold
...
  models.PublicUser:
    properties:
      avatarUrl:
        type: string
      dormBuilding:
        type: string
      id:
        type: string
      name:
        type: string
      trustScore:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: tnswnb@gmail.com
    name: DormMarket Support
  description: REST API สำหรับแพลตฟอร์มตลาดซื้อขายของมือสองในหอ/มหาลัย (DormMarket)
  license:
    name: MIT
  title: DormMarket API
  version: "1.0"
paths:
...
  /api/listings:
    get:
      parameters:
      - description: กรองตามหมวดหมู่
        in: query
        name: categoryId
        type: string
      - description: กรองตามผู้ขาย
        in: query
        name: sellerId
        type: string
      - description: คำค้นหาในชื่อ/รายละเอียด
        in: query
        name: search
        type: string
      - description: ราคาต่ำสุด
        in: query
        name: minPrice
        type: integer
      - description: ราคาสูงสุด
        in: query
        name: maxPrice
        type: integer
      - description: เลื่อนหน้า (pagination offset)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Listing'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: ค้นหา/รายการประกาศ
      tags:
      - listings
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลประกาศ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createListingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Listing'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลงประกาศขายสินค้าใหม่
      tags:
      - listings
  /api/listings/{id}:
    delete:
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ลบสำเร็จ
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลบประกาศ (soft delete, เจ้าของเท่านั้น)
      tags:
      - listings
    get:
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Listing'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: รายละเอียดประกาศ 1 ชิ้น
      tags:
      - listings
    put:
      consumes:
      - application/json
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลประกาศที่แก้ไข
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Listing'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขประกาศ (เจ้าของเท่านั้น)
      tags:
      - listings
...
  /api/listings/{id}/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: ไฟล์รูปภาพ (อัปโหลดได้หลายรูป)
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Listing'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: อัปโหลดรูปสินค้า (เจ้าของเท่านั้น)
      tags:
      - listings
  /api/listings/{id}/status:
    patch:
      consumes:
      - application/json
      description: 'สถานะที่รองรับ: available, reserved, sold'
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      - description: สถานะใหม่
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: เปลี่ยนสถานะประกาศ (เจ้าของเท่านั้น)
      tags:
      - listings
  /api/listings/search-by-image:
    post:
      consumes:
      - multipart/form-data
      description: รับรูป 1 รูป คืนรายการประกาศที่มีรูปคล้ายกันที่สุด (ต้องเปิดใช้งาน
        ML_SERVICE_URL)
      parameters:
      - description: รูปที่จะใช้ค้นหา
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Listing'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "501":
          description: ปิดใช้งาน image search
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: ค้นหาประกาศด้วยรูปภาพ (image similarity search)
      tags:
      - listings
  /api/listings/suggest-price:
    get:
      description: คำนวณจากราคาเฉลี่ยของประกาศอื่นในหมวดหมู่+สภาพเดียวกันที่มีอยู่แล้วในระบบ
      parameters:
      - description: รหัสหมวดหมู่
        in: query
        name: categoryId
        required: true
        type: string
      - description: สภาพสินค้า (new, like_new, good, worn)
        in: query
        name: condition
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: แนะนำราคาสำหรับสินค้า (rule-based)
      tags:
      - listings
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคแล้วตามด้วย JWT token เช่น "Bearer
      eyJhbGc..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"